Updated all icons on site
-Moved icons from Phosphor icons to Font Awesome Icons

This was done in an effort to reduce the amount of data that had to be loaded and unused CSS.

The previous script from phosphor was reading and downloading the CSS for their entire library.

The new icons only load what is actually used.

Also added a transform hover affect on the "Read More" links.